+++ lib/models/response/woosignal_app.dart
@@ -26,6 +26,9 @@ class WooSignalApp {
   String? stripeAccount;
   int? wpLoginEnabled;
   bool? couponEnabled;
+  bool? showProductReviews;
+  bool? showRelatedProducts;
+  bool? showUpsellProducts;
   String? wpLoginBaseUrl;
   String? wpLoginForgotPasswordUrl;
   String? wpLoginWpApiPath;
@@ -100,6 +103,24 @@ class WooSignalApp {
       couponEnabled = true;
     } else {
       couponEnabled = false;
+    }
+    if (json['show_product_reviews'] is int &&
+        json['show_product_reviews'] == 1) {
+      showProductReviews = true;
+    } else {
+      showProductReviews = false;
+    }
+    if (json['show_related_products'] is int &&
+        json['show_related_products'] == 1) {
+      showRelatedProducts = true;
+    } else {
+      showRelatedProducts = false;
+    }
+    if (json['show_upsell_products'] is int &&
+        json['show_upsell_products'] == 1) {
+      showUpsellProducts = true;
+    } else {
+      showUpsellProducts = false;
     }
     wpLoginBaseUrl = json['wp_login_base_url'];
     wpLoginForgotPasswordUrl = json['wp_login_forgot_password_url'];