+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [3.0.5] - 2022-01-29
+
+* Version added to woosignal.dart
+* Flutter format
+* Improvements as per Dart analysis
+
+## [3.0.4] - 2022-01-28
+
+* Add new variables to woosignal_app
+* Update arguments for product review
+* Dependency updates
+
 ## [3.0.3] - 2022-01-03
 
 * Update coupon property types